requirements.txt
requests
flask
mysql-connector-python
python-dotenv